bot/requirements.txt
loguru==0.7.3
numpy==2.2.6
openai==1.66.3
opencv-python==4.11.0.86
pydantic==2.11.5
pymongo==4.12.0
pytest==8.4.0
pytest-cov==6.1.1
python-dotenv==1.0.1
python-telegram-bot[job-queue,webhooks]==22.0
requests==2.32.3
motor==3.7.1
fastapi==0.115.12
hypercorn==0.17.3
